feature: add auth_token to user
Added a temporary auth_token mechanism to the user model to allow
users to temporarily login via a url parameters when using AAF RapidConnect
to log in for SSO. The URL parameter method was needed since we couldn't pass a
cookie or header since we are creating a redirect in POST /auth/jwt. Thus, the
token is only ever used to login within a 30 second timeframe and then invalidated.
Then we take the usual JWT token route for logging in.

Note the user model has also removed password restrictions since these were causing conflictions
to the API calls and weren't needed since we are using Rails has_secure_password mechanism. Moreover,
we shouldn't be exposing password details in the API client.